Rename Value namespace to Data
